% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espolygon.R
\name{create_espolygon}
\alias{create_espolygon}
\title{Build ESPOLYGON From One or More Source SSURGO Databases}
\usage{
create_espolygon(
  dsn,
  esagg,
  filename,
  mupolygon = es_get_mupolygon_layer(),
  query = NA,
  by.dsn = "MUKEY",
  by.agg = "mukey"
)
}
\arguments{
\item{dsn}{character. Path to one or more File Geodatabase, GeoPackage or SpatiaLite databases.}

\item{esagg}{data.frame containing "all" condition aggregation.}

\item{filename}{Output filename (e.g. a new or existing GeoPackage or SQLite file, or any OGR data source that supports appending data)}

\item{mupolygon}{character. Name of mapunit polygon geometry layer; default \code{"MUPOLYGON"} via \code{es_get_mupolygon_layer()}}

\item{query}{Optional: Query to read only a subset of mapunit polygon geometry source in \code{dsn.} Default \code{NA} to read whole layer.}

\item{by.dsn}{Default: \code{"MUKEY"}}

\item{by.agg}{Default: \code{"mukey"}}
}
\value{
If \code{filename} exists after writ
}
\description{
Build ESPOLYGON From One or More Source SSURGO Databases
}
