% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukey.R
\name{create_mukey_lut}
\alias{create_mukey_lut}
\title{Create MUKEY Look-up Table}
\usage{
create_mukey_lut(WHERE = NULL)
}
\arguments{
\item{WHERE}{Optional: SQL where clause targeting \code{legend} or \code{mapunit} level attributes.}
}
\value{
data.frame containing \code{"mukey"} and \code{"areasymbol"} columns
}
\description{
Creates a look-up table of MUKEY and AREASYMBOL, possibly subject to optional \code{WHERE} clause condition.
}
\examples{
\dontrun{
create_mukey_lut(WHERE = "areasymbol = 'CA041'")
}
}
