% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{es_check_site_changed}
\alias{es_check_site_changed}
\alias{es_check_new_assignnment}
\alias{es_check_top_3}
\title{Methods for Comparing Old and New ES Assignments}
\usage{
es_check_site_changed(x, topn = 4L)

es_check_new_assignnment(x, topn = 4L)

es_check_top_3(x)
}
\arguments{
\item{x}{data.frame result from \code{es_compare_old_new()} or similar}

\item{topn}{Number of sites to consider in comparison. Default is the first 4 conditions.}
}
\value{
logical matrix.
}
\description{
\itemize{
\item \code{es_check_site_changed()}: check sites that changed between two data sources (i.e. name and ID are different)
}

\itemize{
\item \code{es_check_new_assignnment()}: visualize new assignments e.g. \code{old == "Not assigned"} -> \code{new == something})
}

\itemize{
\item \code{es_check_top_n()}: tabulate compositional labels (top n classes)
}
}
