% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esagg.R
\name{es_aggregate_ssurgo}
\alias{es_aggregate_ssurgo}
\title{Get SSURGO Ecological Site Aggregation}
\usage{
es_aggregate_ssurgo(x, threshold = -100, ..., not_assigned = NA, fillna = TRUE)
}
\arguments{
\item{x}{character. Vector of one or more soil survey area symbols of interest.}

\item{threshold}{Passed to \code{soilDB::get_SDA_coecoclass()}. Default: \code{0} ensuring that all conditions, regardless of aggeregate component percentage, are included.}

\item{...}{Additional arguments passed to \code{soilDB::get_SDA_coecoclass()} (such as for a custom data source via \code{dsn} argument)}

\item{not_assigned}{character. Value used for conditions not assigned an ecological site. Default: \code{NA}}

\item{fillna}{logical. Fill any \code{NA} values generated by \code{data.table::rbindlist(fill=TRUE)} in "site" ID or "name" fields with value of \code{not_assigned}}
}
\description{
Performs "all" condition aggregation via \code{soilDB::get_SDA_coecoclass()} (called iteratively).
}
\details{
Suitable for generating aggregate ecological site information for the complete SSURGO database.
}
