% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esagg.R
\name{es_calculate_aggregate_sitepct}
\alias{es_calculate_aggregate_sitepct}
\title{Calculate Aggregate Site Percentage}
\usage{
es_calculate_aggregate_sitepct(x, pattern = "pct_r", na.rm = TRUE)
}
\arguments{
\item{x}{a data.frame containing "all" condition aggregation, one row per map unit}

\item{pattern}{Default: \code{"pct_r"}}

\item{na.rm}{Default: \code{TRUE}; use \code{FALSE} to return \code{NA} if any condition is missing a value.}
}
\value{
numeric. Sum of values columns containing \code{"pct_r"} (or other specified \code{pattern})
}
\description{
Calculate aggregate site percentage from an "all" condition aggregation. Derived from sum of columns in \code{x} containing \code{"pct_r"}
}
