% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esagg.R
\name{es_decompose_site_code}
\alias{es_decompose_site_code}
\title{Parse an Ecological Site or Forage Suitability Group ID}
\usage{
es_decompose_site_code(x)
}
\arguments{
\item{x}{character. Vector of site codes to parse. Standard site codes for \code{ecoclasstypename="NRCS Forestland Site"} and
\code{ecoclasstypename="NRCS Rangeland Site"} for \verb{ecoclassref="Ecological Site Description Database} as well as
\code{ecoclasstypename="Forage Suitability Group"} are \code{11} characters in length.}
}
\value{
\emph{data.frame} containing columns \code{"kind"}, \code{"mlra"}, \code{"unit"}, \code{"number"}, and \code{"state"}.
}
\description{
Parse an Ecological Site or Forage Suitability Group ID
}
\examples{
es_decompose_site_code(c("R103XY021MN", "G103XS006MN", "F108XC508IA", NA, "C01948"))
}
