% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esagg.R
\name{es_remove_empty_conditions}
\alias{es_remove_empty_conditions}
\title{Remove Conditions all \code{"Not assigned"} or \code{NA}}
\usage{
es_remove_empty_conditions(x, not_assigned = "Not assigned")
}
\arguments{
\item{x}{a data.frame containing "all" condition aggregation, one row per map unit}

\item{not_assigned}{character. Value of \code{"Not assigned"} condition. Default: \code{"Not assigned"}}
}
\value{
data.frame, with some columns potentially omitted.
}
\description{
Remove Conditions all \code{"Not assigned"} or \code{NA}
}
