% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESAT-package.R
\docType{package}
\name{ESAT-package}
\alias{ESAT}
\alias{ESAT-package}
\title{ESAT: Ecological Site Analytical Tools}
\description{
Provides tools for analysis of the tabular and spatial configuration of Ecological Site Concepts in SSURGO (Soil Survey Geographic Database) products.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/ESAT}
}

}
\keyword{internal}
