% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{es_set_ssurgo_source}
\alias{es_set_ssurgo_source}
\alias{es_set_mupolygon_layer}
\alias{es_get_mupolygon_layer}
\alias{es_set_sapolygon_layer}
\alias{es_get_sapolygon_layer}
\title{Set Path to SSURGO Database}
\usage{
es_set_ssurgo_source(x, mupolygon = NULL, sapolygon = NULL)

es_set_mupolygon_layer(x)

es_get_mupolygon_layer()

es_set_sapolygon_layer(x)

es_get_sapolygon_layer()
}
\arguments{
\item{x}{character. Path to File Geodatabase or GeoPackage/SpatiaLite database.}

\item{mupolygon}{character. name of the MUPOLYGON or SAPOLYGON layer within \code{x}.}

\item{sapolygon}{character. name of the MUPOLYGON or SAPOLYGON layer within \code{x}.}
}
\description{
Either File Geodatabase or GeoPackage/SpatiaLite database supported.
}
